Adding Bevy engine and mentioning Bluetooth
The Bevy engine is a new game engine made in Rust and it's seeing a lot of enthusiasm in the community, I think it should be mentioned here.

These days I needed to write an application to communicate with a Bluetooth device and wanted to write it in Rust. I had to give up because I couldn't find a stable crate to use. More details [here](https://www.reddit.com/r/rust/comments/j81m4z/are_we_bluetooth_yet/).